broke out some functions
